// src/app/core/api/model/adminUserRowDto.ts
/**
 * AGORA API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AdminUserRowDto { 
    id?: string;
    email?: string;
    firstName?: string;
    lastName?: string;
    accountType?: string;
    status?: string;
    phone?: string;
    internalRef?: string;
    notesAdmin?: string;
    exemptions?: Array<string>;
    createdAt?: string;
}