// src/app/core/api/model/updateAdminGroupRequestDto.ts
/**
 * AGORA API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface UpdateAdminGroupRequestDto {
  name: string;
  canViewImmobilier?: boolean;
  canBookImmobilier?: boolean;
  canViewMobilier?: boolean;
  canBookMobilier?: boolean;
  discountType: UpdateAdminGroupRequestDto.DiscountTypeEnum;
  discountValue?: number;
  discountAppliesTo: UpdateAdminGroupRequestDto.DiscountAppliesToEnum;
}
export namespace UpdateAdminGroupRequestDto {
  export const DiscountTypeEnum = {
    None: 'NONE',
    Percentage: 'PERCENTAGE',
    FixedAmount: 'FIXED_AMOUNT',
    FullExempt: 'FULL_EXEMPT',
  } as const;
  export type DiscountTypeEnum = (typeof DiscountTypeEnum)[keyof typeof DiscountTypeEnum];
  export const DiscountAppliesToEnum = {
    All: 'ALL',
    ImmobilierOnly: 'IMMOBILIER_ONLY',
    MobilierOnly: 'MOBILIER_ONLY',
  } as const;
  export type DiscountAppliesToEnum =
    (typeof DiscountAppliesToEnum)[keyof typeof DiscountAppliesToEnum];
}